Rename local extension class
